feat: 修改后端代码结构，增加功能接口
